requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.44.0
playwright-stealth>=1.0.6
trafilatura>=1.12.0
rich>=13.7.0
openai>=1.0.0
pydantic>=2.0.0
ddgs>=0.1.0
PyYAML>=6.0
